Backend/pyproject.toml
[project]
name = "agent-data-analyst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.3",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "replicate>=1.0.7",
    "scipy>=1.16.1",
]
